'set-cookie-cart'

+++ ShopOrder/Utils/FileUploads.cs
@@ -8,9 +8,37 @@ namespace ShopOrder.Utils
 {
     public class FileUploads
     {
-        public static string Upload(HttpServerUtilityBase server, HttpPostedFileBase file)
+        public static string UploadMatHang(HttpServerUtilityBase server, HttpPostedFileBase file)
         {
-            string tempFolder = Path.Combine(server.MapPath("~/Images/Upload/"));
+            return Upload(server, file, FileFolders.MatHang);
+        }
+
+        public static void DeleteMatHang(HttpServerUtilityBase server, string fileName)
+        {
+            Delete(server, fileName, FileFolders.MatHang);
+        }
+        public static string UploadKhachHang(HttpServerUtilityBase server, HttpPostedFileBase file)
+        {
+            return Upload(server, file, FileFolders.KhachHang);
+        }
+
+        public static void DeleteKhachHang(HttpServerUtilityBase server, string fileName)
+        {
+            Delete(server, fileName, FileFolders.KhachHang);
+        }
+        public static string UploadDonHang(HttpServerUtilityBase server, HttpPostedFileBase file)
+        {
+            return Upload(server, file, FileFolders.DonHang);
+        }
+
+        public static void DeleteDonHang(HttpServerUtilityBase server, string fileName)
+        {
+            Delete(server, fileName, FileFolders.DonHang);
+        }
+
+        private static string Upload(HttpServerUtilityBase server, HttpPostedFileBase file, string folders)
+        {
+            string tempFolder = Path.Combine(server.MapPath(string.Format("~/Images/{0}/", folders)));
             if (!Directory.Exists(tempFolder))
             {
                 Directory.CreateDirectory(tempFolder);
@@ -29,8 +57,9 @@ namespace ShopOrder.Utils
                 fileName = fileName + ".jpg";
             }
 
-            string path = Path.Combine(server.MapPath("~/Images/Upload/" + fileName));
-
+            string path = string.Format("~/Images/{0}/{1}", folders, fileName);
+            //path = Path.Combine(path);
+            path = server.MapPath(path);
             if (File.Exists(path)) File.Delete(path);
 
             file.SaveAs(path);
@@ -38,13 +67,20 @@ namespace ShopOrder.Utils
             return fileName;
         }
 
-        public static void Delete(HttpServerUtilityBase server, string fileName)
+        private static void Delete(HttpServerUtilityBase server, string fileName, string folders)
         {
-            string path = Path.Combine(server.MapPath("~/Images/Upload/" + fileName));
+            if (fileName.ToLower().Contains("noimage.jpg") || fileName.ToLower().Contains("no-image.jpg")) return;
+            string path = Path.Combine(string.Format("~/Images/{0}/{1}", folders, fileName));
             if (File.Exists(path))
             {
                 File.Delete(path);
             }
         }
+    }
+
+    public class FileFolders {
+        public const string MatHang = "Items";
+        public const string KhachHang = "Customers";
+        public const string DonHang = "Invoices";
     }
 }
